removes the slime ghostrole

+++ code/modules/mob/living/carbon/slime/items.dm
@@ -125,7 +125,7 @@
 
 /obj/item/docility_serum
 	name = "docility serum"
-	desc = "A potent chemical mix that will nullify a slime's powers, causing it to become docile and tame. This one is meant for adult slimes"
+	desc = "A potent chemical mix that will nullify a slime's powers, causing it to become docile and tame. This one is meant for baby slimes"
 	icon = 'icons/obj/chemical.dmi'
 	icon_state = "bottle-1"
 
@@ -144,40 +144,64 @@
 	if(M.mind)
 		to_chat(user, SPAN_WARNING("The slime is too intelligent to be pacified!"))
 		return ..()
-
 	if(M.is_adult)	
-		var/mob/living/simple_animal/adultslime/pet = new /mob/living/simple_animal/adultslime(M.loc)
-		pet.icon_state = "[M.colour] adult slime"
-		pet.icon_living = "[M.colour] adult slime"
-		pet.icon_dead = "[M.colour] baby slime dead"
-		pet.colour = "[M.colour]"
-		to_chat(user, "You feed the slime the docility serum, removing it's powers and calming it.")
-		qdel(M)
-	
-		var/newname = sanitize(input(user, "Would you like to give the slime a name?", "Name your new pet", "pet slime") as null|text, MAX_NAME_LEN)
+		to_chat(user, SPAN_WARNING("The serum isn't advanced enough to affect adult slimes."))
+		return ..()
 
-		if(!newname)
-			newname = "pet slime"
-		pet.name = newname
-		pet.real_name = newname
-		qdel(src)
-	
-	else
-		var/mob/living/simple_animal/slime/pet = new /mob/living/simple_animal/slime(M.loc)
-		pet.icon_state = "[M.colour] baby slime"
-		pet.icon_living = "[M.colour] baby slime"
-		pet.icon_dead = "[M.colour] baby slime dead"
-		pet.colour = "[M.colour]"
-		to_chat(user, SPAN_NOTICE("You feed the slime the docility serum, removing its powers and calming it."))
-		qdel(M)
+	var/mob/living/simple_animal/slime/pet = new /mob/living/simple_animal/slime(M.loc)
+	pet.icon_state = "[M.colour] baby slime"
+	pet.icon_living = "[M.colour] baby slime"
+	pet.icon_dead = "[M.colour] baby slime dead"
+	pet.colour = "[M.colour]"
+	to_chat(user, SPAN_NOTICE("You feed the slime the docility serum, removing its powers and calming it."))
+	qdel(M)
 
-		var/newname = sanitize(input(user, "Would you like to give the slime a name?", "Name your new pet", "pet slime") as null|text, MAX_NAME_LEN)
+	var/newname = sanitize(input(user, "Would you like to give the slime a name?", "Name your new pet", "pet slime") as null|text, MAX_NAME_LEN)
+	if(!newname)
+		newname = "pet slime"
+	pet.name = newname
+	pet.real_name = newname
+	qdel(src)
 
-		if(!newname)
-			newname = "pet slime"
-		pet.name = newname
-		pet.real_name = newname
-		qdel(src)
+/obj/item/advanced_docility_serum
+	name = "advanced docility serum"
+	desc = "A potent chemical mix that will nullify a slime's powers, causing it to become docile and tame. This one is meant for adult slimes"
+	icon = 'icons/obj/chemical.dmi'
+	icon_state = "bottle-1"
+
+/obj/item/advanced_docility_serum/Initialize() // Better than hardsprited in stuff.
+	var/mutable_appearance/filling = mutable_appearance('icons/obj/reagentfillings.dmi', "[icon_state]-100")
+	filling.color = COLOR_PALE_PINK
+	add_overlay(filling)
+
+/obj/item/advanced_docility_serum/attack(mob/living/carbon/slime/M as mob, mob/user as mob)
+	if(!istype(M, /mob/living/carbon/slime/))//If target is not a slime.
+		to_chat(user, SPAN_WARNING("The docility serum only works on slimes!"))
+		return ..()
+	if(M.stat)
+		to_chat(user, SPAN_WARNING("The slime is dead!"))
+		return ..()
+	if(M.mind)
+		to_chat(user, SPAN_WARNING("The slime is too intelligent to be pacified!"))
+		return ..()
+	if(!M.is_adult)	
+		to_chat(user, SPAN_WARNING("The serum is too advanced to affect baby slimes."))
+		return ..()
+
+	var/mob/living/simple_animal/slime/pet = new /mob/living/simple_animal/slime(M.loc)
+	pet.icon_state = "[M.colour] adult slime"
+	pet.icon_living = "[M.colour] adult slime"
+	pet.icon_dead = "[M.colour] baby slime dead"
+	pet.colour = "[M.colour]"
+	to_chat(user, SPAN_NOTICE("You feed the slime the advanced docility serum, removing its powers and calming it."))
+	qdel(M)
+
+	var/newname = sanitize(input(user, "Would you like to give the slime a name?", "Name your new pet", "pet slime") as null|text, MAX_NAME_LEN)
+	if(!newname)
+		newname = "pet slime"
+	pet.name = newname
+	pet.real_name = newname
+	qdel(src)
 
 /obj/item/slimesteroid
 	name = "slime steroid"
@@ -218,40 +242,6 @@
 	var/mutable_appearance/filling = mutable_appearance('icons/obj/reagentfillings.dmi', "[icon_state]-100")
 	filling.color = COLOR_BLUE
 	add_overlay(filling)
-
-// ghost role slimes & golems
-
-/obj/item/intelligence_serum
-	name = "intelligence serum"
-	desc = "A potent chemical mix that will enhance a slime's intelligence. It must be administered to a baby slime."
-	icon = 'icons/obj/chemical.dmi'
-	icon_state = "bottle-1"
-
-/obj/item/intelligence_serum/Initialize() // Better than hardsprited in stuff.
-	var/mutable_appearance/filling = mutable_appearance('icons/obj/reagentfillings.dmi', "[icon_state]-100")
-	filling.color = COLOR_PALE_PINK
-	add_overlay(filling)
-
-/obj/item/intelligence_serum/attack(mob/living/carbon/slime/M as mob, mob/user as mob)
-	if(!istype(M, /mob/living/carbon/slime)) //If target is not a slime.
-		to_chat(user, SPAN_WARNING("Unfortunately for you, the intelligence serum only works on slimes."))
-		return ..()
-	if(M.is_adult) //Requires a baby slime.
-		to_chat(user, SPAN_WARNING("The intelligence serum requires a younger mind to enhance."))
-		return ..()
-	if(M.stat == DEAD)
-		to_chat(user, SPAN_WARNING("The slime is dead!"))
-		return ..()
-	if(M.mind)
-		to_chat(user, SPAN_WARNING("This slime is already very intelligent!"))
-		return ..()
-	if(M.seeking_player)
-		to_chat(user, SPAN_WARNING("This slime has already been given the intelligence serum."))
-		return ..()
-
-	to_chat(user, "You infuse the intelligence serum into the slime.")
-	M.request_player()
-	qdel(src)
 
 /obj/effect/golemrune
 	anchored = TRUE
